requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argcomplete==3.6.2
asttokens
async-lru==2.0.5
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
Brotli==1.1.0
cachetools==6.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm
contourpy==1.3.2
cycler==0.12.1
dacite==1.7.0
debugpy
decorator
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
einops==0.8.1
exceptiongroup
executing
fiftyone==1.5.2
fiftyone-brain==0.21.2
fiftyone_db==1.1.7
filelock==3.13.1
filetype==1.2.0
fire==0.7.0
Flask==3.1.1
fonttools==4.58.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
glob2==0.7
graphql-core==3.2.6
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
humanize==4.12.3
hydra-core==1.3.2
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.7
imageio==2.37.0
imantics==0.1.12
importlib_metadata
inflate64==1.0.1
iopath==0.1.9
ipykernel
ipython
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jupyter_client
jupyter_core
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.4.0
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
mypy_extensions==1.1.0
narwhals==1.41.0
nest_asyncio
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging
pandas==2.2.3
parso
pathspec==0.12.1
pexpect
pickleshare
pillow==11.0.0
pillow_heif==0.22.0
platformdirs
plotly==6.1.1
portalocker==3.1.1
pprintpp==0.4.0
priority==2.0.0
prompt_toolkit
protobuf==6.31.0
psutil
ptyprocess
pure_eval
py-cpuinfo==9.0.0
py7zr==0.22.0
pybboxes==0.1.6
pybcj==1.0.6
pycocotools==2.0.8
pycryptodomex==3.23.0
pydash==8.0.5
Pygments
pymongo==4.9.2
pyparsing==3.2.3
pyppmd==1.1.1
python-dateutil
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq
pyzstd==0.17.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
roboflow==1.1.64
rtree==1.4.0
s3transfer==0.13.0
safetensors==0.5.3
sahi==0.11.23
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
six
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sse-starlette==0.10.3
sseclient-py==1.8.0
stack_data
starlette==0.46.2
strawberry-graphql==0.270.2
sympy==1.13.3
tabulate==0.9.0
taskgroup==0.2.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.20.3
tomli==2.2.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado
tqdm==4.67.1
traitlets
transformers==4.45.2
triton==2.2.0
typing_extensions
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.3
ultralytics-thop==2.0.14
universal-analytics-python3==1.1.1
urllib3==2.4.0
voxel51-eta==0.14.0
wcwidth
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xmljson==0.2.1
xmltodict==0.14.2
yacs==0.1.8
zipp
